[FIX] user -> admin domain 변경
